macro: fix c_clz() with 64bit literals
If you run c_clz(), and as such _Generic() with string-literals and
evaluate it at *runtime*, any _Generic() selection that is not hit is
still compiled and might produce dead-code. This is fine and is omitted
with -O by gcc, but it still can produce compile-time warnings and errors
if you pass literals. In those cases, we get implicit truncations, which
is really bad.

Add explicit casts, which are no-ops but protect us from errors in those
dead-code paths.

+++ src/bus1/c-macro.h
@@ -487,11 +487,12 @@ extern "C" {
  *
  * Return: Evaluates to an 'int', the number of leading zeroes.
  */
-#define c_clz(_val)                                             \
-        _Generic((_val),                                        \
-                unsigned int: __builtin_clz(_val),              \
-                unsigned long: __builtin_clzl(_val),            \
-                unsigned long long: __builtin_clzll(_val))
+#define c_clz(_val)                                                     \
+        _Generic((_val),                                                \
+                /* cast to avoid compile-warns/errs in dead-code */     \
+                unsigned int: __builtin_clz((unsigned int)(_val)),      \
+                unsigned long: __builtin_clzl((unsigned long)(_val)),   \
+                unsigned long long: __builtin_clzll((unsigned long long)(_val)))
 
 /**
  * c_align_to() - align value to
